requirements.txt
streamlit
pandas
numpy
scikit-learn
plotly
joblib
streamlit-community-navigation-bar==4.4.1
google-cloud-storage
google-auth
pymysql